refactor(app): split parameter-widget-renderer into per-type components
Extract 3 shared hooks (useParamActions, useSeedQueryOptions,
useCascadingClear) and 8 per-type components into parameters/ subfolder.
Main renderer becomes a thin dispatcher. Split monolithic test file into
10 per-type test files for better readability.

Closes #262
